requirements.txt
rembg==2.0.50
Pillow>=9.0.0
